refactor(audit): fold tool metadata audit into the mcp_server subject
The tool metadata audit was modelled as a detached subject: its own
subject type (mcp_server_tool_metadata), its own action slug
(mcp-server-tool-metadata:update), its own URN type, and its own outbox
event. But the thing being audited is the MCP server — a batch write is
just another mutation on it, no different in kind from mcp-server:update.

Treat it that way. The action becomes mcp-server:update-tool-metadata, a
feature-specific verb on the existing mcp_server subject, and the entry
is now identified by urn.McpServer, so a user reading a server's audit
history sees its tool metadata changes alongside its other mutations.

That makes the dedicated URN type and the mcp_server_tool_metadata_event
outbox def dead; both are removed (the outbox event never shipped). The
snapshot fields are renamed to ToolMetadata{Before,After} to keep them
honest about carrying the tool collection rather than the server view.
