+++ server/internal/outbox/events/catalog_gen.yaml
@@ -1246,60 +1246,6 @@ webhooks:
                                 - subject_type
                             type: object
                 required: true
-    audit_log.mcp_server_tool_metadata_event_v1:
-        post:
-            description: Emitted when changes to MCP server tool metadata are made
-            operationId: audit_log.mcp_server_tool_metadata_event_v1
-            requestBody:
-                content:
-                    application/json:
-                        schema:
-                            additionalProperties: false
-                            properties:
-                                action:
-                                    type: string
-                                actor_display_name:
-                                    type: string
-                                actor_id:
-                                    type: string
-                                actor_slug:
-                                    type: string
-                                actor_type:
-                                    type: string
-                                after_snapshot:
-                                    additionalProperties: true
-                                before_snapshot:
-                                    additionalProperties: true
-                                id:
-                                    format: uuid
-                                    type: string
-                                metadata:
-                                    additionalProperties: true
-                                organization_id:
-                                    type: string
-                                project_id:
-                                    format: uuid
-                                    type:
-                                        - string
-                                        - "null"
-                                subject_display_name:
-                                    type: string
-                                subject_id:
-                                    type: string
-                                subject_slug:
-                                    type: string
-                                subject_type:
-                                    type: string
-                            required:
-                                - id
-                                - organization_id
-                                - actor_id
-                                - actor_type
-                                - action
-                                - subject_id
-                                - subject_type
-                            type: object
-                required: true
     audit_log.model_provider_key_event_v1:
         post:
             description: Emitted when changes to customer model provider keys are made
